change data format

+++ server.v
@@ -4,11 +4,10 @@ import vweb
 import sqlite
 import json
 import strings
-import encoding.base64
-import net.urllib
 
 const (
 	port = 8082
+	BASE_TABLE = 'custom'
 )
 
 pub struct App {
@@ -54,6 +53,11 @@ struct Column{
 	mut :
 		sql_name string [ignore]
 }
+struct Data{
+	id string
+	table string
+	data string
+}
 
 fn main() {
 	vweb.run<App>(port)
@@ -66,9 +70,24 @@ pub fn (app mut App) index() {
 fn insert_custom_table(table Table) {
 	detail:=json.encode(table.detail)
 	db := sqlite.connect('test.db')
-	_ ,code:= db.exec("insert into custom (name,sql_name,detail,version) values ('$table.name','$table.sql_name','$detail',$table.version)")
-	println('insert into custom == $code')
-	db.exec('DROP TABLE IF EXISTS $table.sql_name')
+	//check if the table exists
+	rows,_:=db.exec("select * from $BASE_TABLE  where sql_name=='$table.sql_name'")
+	println('check if the table exists == $rows')
+	if rows.len>0 {
+		_,code:=db.exec("update $BASE_TABLE set name='$table.name',sql_name='$table.sql_name',detail='$detail',version=$table.version where id=${rows[0].vals[0]}")
+		db.exec('DROP TABLE IF EXISTS $table.sql_name')
+		if code ==101 {
+			create_table(table)
+		}
+	}else{
+		_,code:=db.exec("insert into $BASE_TABLE (name,sql_name,detail,version) values ('$table.name','$table.sql_name','$detail',$table.version)")
+		if code ==101 {
+			create_table(table)
+		}
+	}
+}
+
+pub fn create_table(table Table){
 	mut sql := strings.Builder{}
 	sql.write('CREATE TABLE IF NOT EXISTS "$table.sql_name" (id integer PRIMARY KEY autoincrement,') 
 	for col in table.detail {
@@ -76,6 +95,7 @@ fn insert_custom_table(table Table) {
 	}
 	sql.go_back(1)
 	sql.write(')')
+	db := sqlite.connect('test.db')
 	_, sqlcode := db.exec(sql.str())
 	println('create table == $sql.str()  ==  $sqlcode')
 }
@@ -105,45 +125,62 @@ pub fn (app mut App) tables() {
 
 pub fn (app mut App) add_table() {
 	app.vweb.add_header('Access-Control-Allow-Origin' , '*')
-	println('add_table start =='+app.vweb.form['data'] +' == '+app.vweb.req.data)
-	mut table:=json.decode(Table,app.vweb.form['data']) or{
+	println('add_table start == '+app.vweb.req.data)
+	mut table:=json.decode(Table,app.vweb.req.data) or{
 		app.vweb.text('shibai')
 		return
 	}
-	table.sql_name=base64.encode(table.name)
+	// table.sql_name=base64.encode(table.name)
+	table.sql_name=table.name
 	for i,col in table.detail {
-		table.detail[i].sql_name=base64.encode(col.name)
+		// table.detail[i].sql_name=base64.encode(col.name)
+		table.detail[i].sql_name=col.name
 	}
+	println('encode == '+json.encode(table))
 	insert_custom_table(table)
 	app.vweb.text('jieshu')
 }
 
 pub fn (app mut App) delete_table() {
 	app.vweb.add_header('Access-Control-Allow-Origin' , '*')
-	table:=app.vweb.form['table']
+	table:=app.vweb.req.data
+	println('delete table == $table')
 	db := sqlite.connect('test.db')
-	db.exec('delete from custom where sql_name="$table"')
+	_,code:=db.exec('delete from custom where sql_name="$table"')
 	db.exec('drop table "$table";')
+	println('delete from custom where sql_name="$table" == $code')
 	app.vweb.text('jieshu')
 }
 
 pub fn (app mut App) list_data() {
 	app.vweb.add_header('Access-Control-Allow-Origin' , '*')
-	table_name := app.vweb.form['table']
+	table_name := app.vweb.req.data
 	db := sqlite.connect('test.db')
-	rows, _ := db.exec('select * from "$table_name"')
-	println(rows)
-	// mut infos:=[]map[string]string
-	// for row in rows {
-		
-	// }
+	rows, code := db.exec('select * from "$table_name"')
+	println('select * from "$table_name" == $code')
 	app.vweb.json(json.encode(rows))
 }
 
 pub fn (app mut App) add_data() {
 	app.vweb.add_header('Access-Control-Allow-Origin' , '*')
-	table_name:=app.vweb.form['table']
-	mut data_str := app.vweb.form['data']
+	mut data_str := app.vweb.req.data
+	println(data_str)
+	// data:=json.decode(Data,data_str) or{
+	// 	app.vweb.text('shibai')
+	// 	return
+	// }
+	// println(json.encode(data))
+	start :=data_str.index('"table":"') or {
+		return
+	}
+	end:=data_str.index('",') or{
+		return 
+	}
+	table_name:=data_str.substr(start+9,end)
+	data_start :=data_str.index('"data":') or {
+		return
+	}
+	data_str=data_str.substr(data_start+7,data_str.len-1)
 	mut sqlstr:=strings.Builder{}
 	sqlstr.write('insert into "$table_name" (')
 	data_str = data_str.trim_left('{')
@@ -166,14 +203,25 @@ pub fn (app mut App) add_data() {
 	db := sqlite.connect('test.db')
 	_,code:=db.exec(sqlstr.str())
 	println(sqlstr.str() +' == $code')
-
 	app.vweb.text('jieshu')
 }
 
 pub fn (app mut App) delete_data() {
 	app.vweb.add_header('Access-Control-Allow-Origin' , '*')
-	table:=app.vweb.form['table']
-	id:=app.vweb.form['id']
+	data_str:=app.vweb.req.data
+	println(data_str)
+	mut table:=''
+	mut id:=''
+	ss:=data_str.split(',')
+	for s in ss {
+		ts:=s.split(':')
+		if ts[0].contains('id') {
+			id=ts[1]
+		}
+		if ts[0].contains('table') {
+			table=ts[1]
+		}
+	}
 	db := sqlite.connect('test.db')
 	_,code:=db.exec('delete from "$table" where id=$id')
 	println('delete from "$table" where id=$id == $code')
